requirements.txt
pydantic==1.7.3
fastapi==0.63.0
Sqlalchemy==1.3.22
sqltap==0.3.11
alembic==1.5.3
psycopg2-binary==2.8.6
http-status-code==0.0.8
passlib[bcrypt]==1.7.4
python-i18n[YAML]==0.3.9
coverage==5.4
PyJWT==2.0.1
email-validator==1.1.2
python-multipart==0.0.5

